+++ HISTORY.md
@@ -1,6 +1,18 @@
 Revision History
 ======
 
+Revision 1.0.6
+------
+
+- Added python wrapping support for _some_ of the functions. Those which don't have any `allocatable, intent(out)` parameters can be compiled into a python module my calling `make F90=gfortran` in the new `./wrap/` directory. It generates a shared library `_bcs.so` and a python module `bcs.py`. In order to run correctly, it needs [f90wrap](https://github.com/jameskermode/f90wrap) installed.
+- Changed the function `file_exists` to use `inquire` instead of attempting an open since that creates the missing file on some compilers.
+- Added a primitive control for the oscillation problem where the solution jumps between two `l0` norms.
+
+Revision 1.0.5
+------
+
+- Worked around a bug in `gfortran` where assignment of a variable's value inside of an associate caused spurious errors with subtraction not being performed correctly. We have yet post the bug to `gfortran`'s bug list.
+
 Revision 1.0.4
 ------